// src/lib/api/generated/models/packMediaControllerUploadBodyRole.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type PackMediaControllerUploadBodyRole = typeof PackMediaControllerUploadBodyRole[keyof typeof PackMediaControllerUploadBodyRole];


export const PackMediaControllerUploadBodyRole = {
  COVER: 'COVER',
  GALLERY: 'GALLERY',
} as const;